Removed most of game manager
