--- Assets/BoardVisualizer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d919537774f73b142809538a3b0ccac6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
